Add utils file and move data cleaning into it
