+++ src/Components/ArtistAlbums.js
@@ -1,12 +1,12 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
 import '../Styles/ArtistAlbums.scss'
+import PropTypes from 'prop-types'
+import { sortAlbums } from '../utilities'
 
 const ArtistAlbums = (props) => {
   const albumList = () => {
-    const albums = props.albums.sort((a, b) => {
-      return a.releaseYear - b.releaseYear
-    })
+    const albums = sortAlbums(props.albums)
     
     return albums.map(album => {
       return (
@@ -29,3 +29,7 @@ const ArtistAlbums = (props) => {
 }
 
 export default ArtistAlbums
+
+ArtistAlbums.propTypes = {
+  name: PropTypes.arrayOf(PropTypes.objects).isRequired,
+}
